add ext pool
